+++ src/main/java/com/postgraduate/domain/auth/application/usecase/kakao/KakaoSignInUseCase.java
@@ -2,7 +2,7 @@ package com.postgraduate.domain.auth.application.usecase.kakao;
 
 import com.postgraduate.domain.auth.application.dto.AuthUserResponse;
 import com.postgraduate.domain.auth.application.dto.KakaoUserInfoResponse;
-import com.postgraduate.domain.auth.application.dto.SignUpRequest;
+import com.postgraduate.domain.auth.application.dto.req.SignUpRequest;
 import com.postgraduate.domain.auth.application.mapper.AuthMapper;
 import com.postgraduate.domain.user.application.usecase.UserCheckUseCase;
 import com.postgraduate.domain.user.domain.entity.User;
