+++ src/main/java/com/postgraduate/domain/auth/presentation/AuthController.java
@@ -1,11 +1,11 @@
 package com.postgraduate.domain.auth.presentation;
 
 import com.postgraduate.domain.auth.application.dto.AuthUserResponse;
-import com.postgraduate.domain.auth.application.dto.SignUpRequest;
+import com.postgraduate.domain.auth.application.dto.req.SignUpRequest;
 import com.postgraduate.domain.auth.application.usecase.jwt.JwtUseCase;
 import com.postgraduate.domain.auth.application.usecase.kakao.KakaoSignInUseCase;
 import com.postgraduate.domain.auth.application.dto.JwtTokenResponse;
-import com.postgraduate.domain.auth.application.dto.KakaoLoginRequest;
+import com.postgraduate.domain.auth.application.dto.req.KakaoLoginRequest;
 import com.postgraduate.global.auth.AuthDetails;
 import com.postgraduate.global.dto.ResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
