+++ src/main/java/com/postgraduate/domain/user/application/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.postgraduate.domain.user.application.mapper;
 
-import com.postgraduate.domain.auth.application.dto.SignUpRequest;
+import com.postgraduate.domain.auth.application.dto.req.SignUpRequest;
 import com.postgraduate.domain.user.application.dto.res.UserInfoResponse;
 import com.postgraduate.domain.user.domain.entity.User;
